refactor: api client context
